requirements.txt
lxml==5.3.2
python-docx==0.8.11
typing_extensions==4.13.2
watchdog==3.0.0
Pillow==10.2.0
